requirements.txt
asgiref==3.9.1
Django==5.2.4
django-autoslug==1.9.9
django-birthday==0.1.4
django-filter==25.1
django-parler==2.3
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
Markdown==3.9
phonenumbers==9.0.14
pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
virtualenv==20.34.0
